Update
Fix the major matching issue. The algorithm is able to realign the data set based on the warping path. The performance looks good. However, the accuracy is not as great as human decision compared with human eyes.

Future plan: Based on this algorithm, try to find a way to detect the abnormal data set